% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{locs1}
\alias{locs1}
\title{Master Locations for Smoking Simulation}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 11 rows and 3 columns.}
\usage{
locs1
}
\description{
This contains a few different representative locations
of where I work, live, and shop. It's to be used as
baseline for simulating a realistic smoking profile of
a candidate smoker. I should be able to add a few more
but this works for now.
}
\keyword{datasets}

