% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randTimes.R
\name{randTimes}
\alias{randTimes}
\title{Return a Vector of N Times Within a Day}
\usage{
randTimes(N, st = "2016-06-14 07:30:00", et = "2016-06-14 22:30:00")
}
\arguments{
\item{N}{- number of events to return}

\item{st}{- starting time in POSIX form: YYYY-MM-DD HH:MM:SS}

\item{st}{- ending time in POSIX form: YYYY-MM-DD HH:MM:SS}
}
\description{
The goal of this is to return a random sample of times
between two known end points.
This comes from a stack overflow post: http://goo.gl/LDHqSF;
the answer and code comes from Dirk Edelbuettel, but I've modified
it slightly to deal with posix times input
}
\examples{
randTimes(5)
}

