% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tnorm.R
\name{tnorm}
\alias{tnorm}
\title{Truncated Normal Distribution}
\source{
Code written by Jim Clark, Duke University, Durham, NC, USA
}
\usage{
tnorm(n, lo, hi, mu, sig)
}
\arguments{
\item{n}{- integer specifying the number of draws from the distribition to
return}

\item{lo}{- the lower bound specified for the distribution}

\item{hi}{- the upper bound specified for the distribution}

\item{mu}{- the mean of the distribution}

\item{sig}{- the standard deviation of the distribution}
}
\description{
Random generation for a truncated normal distribution for \code{n}
random deviates, with lower bound equal to \code{lo}, upper bound equal
to \code{hi}, mean equal \code{mu}, standard deviation equal to \code{sd}.
}
\examples{
tnorm(5, 2, 7, 5, 1)
}

